Moved BackgroundEventProcessor directly into PrototypeAsyncConsumer
